requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
cached-property==1.5.2
celery==5.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.19
django-cleanup==8.0.0
django-compat==1.0.15
django-import-export==3.2.0
django-rq==2.8.1
django-treebeard==4.6.0
et-xmlfile==1.1.0
gunicorn==20.1.0
importlib-metadata==6.6.0
kombu==5.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
prompt-toolkit==3.0.38
psycopg2==2.9.6
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
rq==1.14.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
TgCrypto==1.2.5
typing_extensions==4.6.1
tzdata==2023.3
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
